% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fileview_annotations.R
\name{add_fileview_annotations}
\alias{add_fileview_annotations}
\title{Add new annotations to a file view}
\usage{
add_fileview_annotations(data, fileview_id)
}
\arguments{
\item{data}{Data frame of a file view with additional columns of annotations}

\item{fileview_id}{Synapse ID of the file view}
}
\description{
File annotations can be updated by querying a file view and editing the
annotation columns, however adding *new* annotations is less straightforward.
`add_fileview_annotations()` streamlines this process. Users can query a file
view and add any columns they desire; `add_fileview_annotations()` will
detect the column type, create a new column on Synapse, add the column to the
file view schema, and then store the user-supplied annotations.
}
\examples{
\dontrun{
library("synapser")
synLogin()

fv_id <- "syn17020234"
fv <- synTableQuery(paste0("SELECT * FROM ", fv_id))
fv_data <- as.data.frame(fv)

fv_data$new_annotation_1 <- "value1"
fv_data$new_annotation_2 <- 5
fv_data$new_annotation_3 <- factor("x", "x")
fv_data$new_annotation_4 <- c(TRUE, FALSE)

add_fileview_annotations(fv_data, fv_id)
}
}
