% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{add_missing_placeholder}
\alias{add_missing_placeholder}
\title{Fill in missing values with default or specified string.}
\usage{
add_missing_placeholder(df, placeholder = "Not Annotated",
  replace_keys = NULL)
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{replace_keys}{}
}
\description{
Fill in missing values with default or specified string.
}
