% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{add_queryview_column}
\alias{add_queryview_column}
\title{Shortcut function that wraps `build_tablequery` and `get_tablequery_url`
to add a new column with link to view table query results}
\usage{
add_queryview_column(df, format = c("markdown", "html", "raw"))
}
\arguments{
\item{df}{Data frame to add a column to}

\item{format}{Output format (one of `"markdown"`, `"html"`, or `"raw"`)}
}
\description{
Shortcut function that wraps `build_tablequery` and `get_tablequery_url`
to add a new column with link to view table query results
}
