% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_datatable}
\alias{as_datatable}
\title{Convert and format data frame as datatable widget.}
\usage{
as_datatable(df, cols_as_code = c())
}
\arguments{
\item{df}{Data frame}

\item{cols_as_code}{Columns to format as code (i.e. with a fixed width font)}
}
\description{
Convert and format data frame as datatable widget.
}
