% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_wiki_markdown}
\alias{as_wiki_markdown}
\title{Format data frame as markdown table}
\usage{
as_wiki_markdown(df, cols_as_code = c())
}
\arguments{
\item{cols_as_code}{}
}
\description{
Format data frame as markdown table
}
