% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{augment_values}
\alias{augment_values}
\title{Augment values in one column with the value of another column.}
\usage{
augment_values(df, augment_keys)
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{augment_keys}{list mapping target columns (list names) for which to
prepend values of corresponding meta columns (list values)}
}
\description{
Augment values in one column with the value of another column.
}
