% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{build_tablequery}
\alias{build_tablequery}
\title{Construct a Synapse SQL-style table query using names/values of data frame
column(s) to compose 'WHERE' clauses}
\usage{
build_tablequery(table_id, ...)
}
\arguments{
\item{...}{}
}
\description{
Construct a Synapse SQL-style table query using names/values of data frame
column(s) to compose 'WHERE' clauses
}
\examples{
\dontrun{
test_df \%>\%
    rowwise() \%>\%
    mutate(query = build_tablequery(table_id, assay))
}
}
