% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{count_values}
\alias{count_values}
\title{Count distinct values within each group for specified keys.}
\usage{
count_values(df, group_keys, count_keys)
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{group_keys}{character vector of keys by which to group}

\item{count_keys}{character vector of keys for which to count files}
}
\description{
Count distinct values within each group for specified keys.
}
