% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{create_synapse_links}
\alias{create_synapse_links}
\title{Convert values to named Synapse URLs based on corresponding Synapse IDs.}
\usage{
create_synapse_links(df, link_keys)
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{link_keys}{list mapping target columns (list names) for which to
construct links from Synapse IDs in corresponding ID columns (list values)}
}
\description{
Convert values to named Synapse URLs based on corresponding Synapse IDs.
}
