% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_by_key}
\alias{filter_by_key}
\title{Remove rows with missing or bad values for specified keys.}
\usage{
filter_by_key(df, filter_keys, bad_values = c("null", "Not Applicable"))
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{filter_keys}{character vector of keys for which to check values}

\item{bad_values}{character vector specifying bad values to exclude}
}
\description{
Remove rows with missing or bad values for specified keys.
}
