% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{generate_key_description}
\alias{generate_key_description}
\title{Generate key description table}
\usage{
generate_key_description(table)
}
\arguments{
\item{table}{A table of annotation key/value descriptions (e.g. one produced
by \code{\link[=get_synapse_annotations]{get_synapse_annotations()}})}
}
\value{
A dataframe containing the description of each annotation key in the
table
}
\description{
Generate key description table
}
\examples{
\dontrun{
dat <- get_synapse_annotations()
generate_key_description(dat)
}
}
