% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{generate_manifest_template}
\alias{generate_manifest_template}
\title{Generate manifest template}
\usage{
generate_manifest_template(keys = NULL)
}
\arguments{
\item{keys}{A character vector of annotation keys.}
}
\value{
A 0-row data frame with column names corresponding to those of a
  Synapse manifest.
}
\description{
Generate manifest template
}
\examples{
generate_manifest_template()

my_keys <- c("study", "assay", "tissue")
generate_manifest_template(my_keys)
}
