% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{generate_value_description}
\alias{generate_value_description}
\title{Generate value description table}
\usage{
generate_value_description(table)
}
\arguments{
\item{table}{A table of annotation key/value descriptions (e.g. one produced
by \code{\link[=get_synapse_annotations]{get_synapse_annotations()}})}
}
\value{
A dataframe containing the description of each annotation value in
the table
}
\description{
Generates a table that describes all of the values an annotation key may take
}
\examples{
\dontrun{
dat <- get_synapse_annotations()
generate_value_description(dat)
}
}
