% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get_annotation_keys}
\alias{get_annotation_keys}
\title{Get all keys associated with a module or set of modules}
\usage{
get_annotation_keys(module, source)
}
\arguments{
\item{module}{A module or modules to retrieve keys from. If missing, all keys
will be returned}

\item{source}{A data frame (such as that returned by
\code{\link[=get_synapse_annotations]{get_synapse_annotations()}}) containing information about annotations.
Should have a \code{module} column and \code{key} column}
}
\value{
A character vector of annotation keys from the specified module(s)
}
\description{
Get all keys associated with a module or set of modules
}
\examples{
\dontrun{
get_keys_from_module("sageCommunity")
get_keys_from_module(c("sageCommunity", "neuro"))
}
}
