% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get_synapse_annotations}
\alias{get_synapse_annotations}
\title{Get Synapse annotations}
\usage{
get_synapse_annotations()
}
\value{
A data frame containing all annotation keys, descriptions, column
  types, maximum sizes, values, value descriptions, sources, and the name of
  the annotation's parent module
}
\description{
Download current annotation values from Synapse and provide them as a data
frame.
}
\examples{
\dontrun{
synLogin()
dat <- get_synapse_annotations()
}
}
