% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_helpers.R
\name{get_table_df}
\alias{get_table_df}
\title{Collect all rows and columns from a Synapse table and return as values
in a data frame.}
\usage{
get_table_df(table_id, cache = FALSE)
}
\arguments{
\item{table_id}{Synapse ID of table}

\item{cache}{Is the table cached? Defaults to `FALSE`}
}
\value{
data frame with table values
}
\description{
Collect all rows and columns from a Synapse table and return as values
in a data frame.
}
