% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_tablequery_url}
\alias{get_tablequery_url}
\title{Construct a URL to view the results of a Synapse table query}
\usage{
get_tablequery_url(table_id, query_string)
}
\arguments{
\item{query_string}{}
}
\description{
Construct a URL to view the results of a Synapse table query
}
