% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortia_internal_tracking.R
\name{get_view}
\alias{get_view}
\title{Get Synapse view}
\usage{
get_view(fileview_id)
}
\arguments{
\item{fileview_id}{A synId c().}
}
\value{
A tibble.
}
\description{
This function can accept a synId corresponding to a fileview or table.
}
\examples{
\dontrun{
fv <- get_view("syn18691012")
table <- get_view("syn20555115")
}

}
