% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{list_values}
\alias{list_values}
\title{List distinct values within each group for specified keys}
\usage{
list_values(df, group_keys, list_keys, list_format = c("html", "csv"))
}
\arguments{
\item{df}{data frame nominally representing a Synapse table or file view}

\item{group_keys}{character vector of keys by which to group}

\item{list_keys}{character vector of keys for which to list values}
}
\description{
List distinct values within each group for specified keys
}
