% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortia_internal_tracking.R
\name{mod_table}
\alias{mod_table}
\title{Bind new files to the existing table, notes deleted files and file/annotation modifications or checks only for file/annotation modifications}
\usage{
mod_table(fv, table, fileview_id = c())
}
\arguments{
\item{fv}{A tibble. A fileview imported with get_v.}

\item{table}{A tibble. A table imported with get_v.}

\item{fileview_id}{A synId c().}
}
\value{
A tibble with notation added to the notes column. ROW_ID and ROW_VERSION remain from existing table to enable storage to Synapse.
}
\description{
Bind new files to the existing table, notes deleted files and file/annotation modifications or checks only for file/annotation modifications
}
\examples{
mod_table(fv = get_v("synId"), table = get_v("synId"))
}
