% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_file_counts_by_annotationkey}
\alias{plot_file_counts_by_annotationkey}
\title{Plot the breakdown of files from a file view according to distinct values
within each specified annotation key (i.e., category).}
\usage{
plot_file_counts_by_annotationkey(view_df, annotation_keys,
  replace_missing = "Not Annotated", chart_height = NULL)
}
\arguments{
\item{view_df}{File view data frame}

\item{annotation_keys}{Annotation keys to group by}

\item{replace_missing}{String to use for missing annotation values (defaults to "Not Annotated").}

\item{chart_height}{Height of the chart in pixels (optional, defaults to automatic sizing).}
}
\description{
Plot the breakdown of files from a file view according to distinct values
within each specified annotation key (i.e., category).
}
\examples{
\dontrun{
plot_keys <- list(assay = "Assay", tumorType = "Tumor Type",
                  diagnosis = "Diagnosis", species = "Species",
                  organ = "Organ", tissue = "Tissue",
                  dataType = "Data Type", study = "Study")
plot_file_counts_by_annotationkey(fileview_df, plot_keys)
}
}
