% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_helpers.R
\name{save_chart}
\alias{save_chart}
\title{Save a static or dynamic chart to a file and store in Synapse.}
\usage{
save_chart(parent_id, chart_filename, plot_object, static = FALSE)
}
\arguments{
\item{parent_id}{Synapse ID of the parent, i.e. location where the chart will
be stored}

\item{chart_filename}{File name under which to save the chart}

\item{plot_object}{Plot object to save}

\item{static}{Is the plot static? Defaults to `FALSE`}
}
\description{
Save a static or dynamic chart to a file and store in Synapse.
}
