% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_helpers.R
\name{save_table}
\alias{save_table}
\title{Create a new table in Synapse within a specified project or update an
existing table.}
\usage{
save_table(project_id, table_name, table_df)
}
\arguments{
\item{project_id}{string containing the Synapse ID for the target project}

\item{table_name}{string containing the name of the table to be stored in
Synapse}

\item{table_df}{data frame containing values to be stored in Synapse table}
}
\value{
Synapse `Table` object for stored table data
}
\description{
This is a convenience/utility function to simplify several operations within
the Synapse R client for creating or updating tables.
}
