% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortia_internal_tracking.R
\name{store}
\alias{store}
\title{Store updated table to Synapse}
\usage{
store(table_to_store, fileview_id)
}
\arguments{
\item{table_to_store}{A tibble. Output from mod_table()}

\item{fileview_id}{A synId c().}
}
\value{
A Synapse client object with storage metadata.
}
\description{
Store updated table to Synapse
}
