% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summarize_by_annotationkey}
\alias{summarize_by_annotationkey}
\title{Count files in a Synapse file view, grouped by annotation keys.}
\usage{
summarize_by_annotationkey(view_df, annotation_keys, table_id,
  synproject_key = NULL, count_cols = NULL, list_cols = NULL,
  augment_keys = NULL, link_keys = NULL, filter_missing = TRUE,
  queryformat = "html", list_format = "html")
}
\arguments{
\item{view_df}{data frame representing Synapse file view}

\item{annotation_keys}{character vector of grouping keys}

\item{table_id}{string representing Synapse ID for file view}

\item{synproject_key}{if 'projectId' is in `annotation_keys`, a string
indicating a corresponding key to rename the column}

\item{count_cols}{character vector of keys for which to count files}

\item{list_cols}{character vector of keys for which to list values}

\item{augment_keys}{list mapping target columns (list names) for which to
prepend values of corresponding meta columns (list values)}

\item{link_keys}{list mapping target columns (list names) for which to
construct links from Synapse IDs in corresponding ID columns (list values)}

\item{filter_missing}{remove records with missing annotation values}

\item{queryformat}{Query format. Defaults to `"html"`}

\item{list_format}{Format of keys}
}
\description{
Count files in a Synapse file view, grouped by annotation keys.
}
\examples{
\dontrun{
group_keys <- c("assay", "tumorType")
list_cols <- "study"
augment_keys <- list(study = "Center Name")
link_keys <- list(study = "Study")
fileview_df \%>\%
   summarize_by_annotationkey(
       annotation_keys = group_keys,
       table_id = master_fileview_id,
       count_cols = count_cols,
       list_cols = list_cols,
       augment_keys = augment_keys,
       link_keys = link_keys
   )
}
}
