% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{write_manifest}
\alias{write_manifest}
\title{Write manifest to file}
\usage{
write_manifest(sheets, file = "annotations_manifest.xlsx")
}
\arguments{
\item{sheets}{Sheets to be written into the xlsx file. Can be one data frame,
or a list of data frames.}

\item{file}{The file to be written to.}
}
\description{
Write manifest to file
}
