% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_wiki_markdown}
\alias{as_wiki_markdown}
\title{Format data frame as markdown table}
\usage{
as_wiki_markdown(df, cols_as_code = c())
}
\arguments{
\item{df}{Data frame}

\item{cols_as_code}{Columns to format as code (i.e. with a fixed width font)}
}
\description{
Format data frame as markdown table
}
