% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortia_internal_tracking.R
\name{check_version}
\alias{check_version}
\title{Compare the fileview and table etag to note file metadata or file itself has been 
modified. This would indicate a user needs to download a new version.}
\usage{
check_version(fv, table)
}
\arguments{
\item{fv}{A tibble. A fileview imported with [get_view()].}

\item{table}{A tibble. A table imported with [get_view()].}
}
\value{
A string of synIds where file has changed.
}
\description{
Compare the fileview and table etag to note file metadata or file itself has been 
modified. This would indicate a user needs to download a new version.
}
\examples{
check_version(fv = get_view("synId"), table = get_view("synId"))
}
