% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{format_summarytable_columns}
\alias{format_summarytable_columns}
\title{Prettify column names based on annotation keys.}
\usage{
format_summarytable_columns(df, facet_cols = c())
}
\arguments{
\item{facet_cols}{}
}
\description{
Prettify column names based on annotation keys.
}
