% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consortia_internal_tracking.R
\name{get_view}
\alias{get_view}
\title{Get Synapse view - fileview or table}
\usage{
get_view(fileview_id)
}
\arguments{
\item{fileview_id}{A synId c().}
}
\value{
A tibble.
}
\description{
Get Synapse view - fileview or table
}
\examples{
fv <- get_view("syn18691012")
table <- get_view("syn20555115")
}
