% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_helpers.R
\name{save_chart}
\alias{save_chart}
\title{Save a static or dynamic chart to a file and store in Synapse.}
\usage{
save_chart(parent_id, chart_filename, plot_object, static = FALSE)
}
\arguments{
\item{static}{}
}
\description{
Save a static or dynamic chart to a file and store in Synapse.
}
